seperate runtime_api definition from the full and light implementations.
This is done as a pre-requisite for adding more stateless runtime api calls used in preprocessor.
Those calls will be implemented and replaced in next commits
