+++ domains/client/block-preprocessor/src/xdm_verifier.rs
@@ -1,4 +1,4 @@
-use crate::state_root_extractor::StateRootExtractor;
+use crate::runtime_api::StateRootExtractor;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::{Error, HeaderBackend};
 use sp_domains::ExecutorApi;
@@ -78,6 +78,7 @@ where
 /// Returns Error when required calls to fetch header info fails.
 pub fn verify_xdm_with_primary_chain_client<PClient, PBlock, SBlock, SRE>(
     primary_chain_client: &Arc<PClient>,
+    at: SBlock::Hash,
     state_root_extractor: &SRE,
     extrinsic: &SBlock::Extrinsic,
 ) -> Result<bool, Error>
@@ -90,7 +91,6 @@ where
     PBlock::Hash: From<SBlock::Hash>,
     SRE: StateRootExtractor<SBlock>,
 {
-    let at = state_root_extractor.block_hash();
     if let Ok(state_roots) = state_root_extractor.extract_state_roots(at, extrinsic) {
         // verify system domain state root
         let best_hash = primary_chain_client.info().best_hash;
