+++ domains/client/domain-executor/src/system_bundle_processor.rs
@@ -2,7 +2,7 @@ use crate::domain_block_processor::{DomainBlockProcessor, PendingPrimaryBlocks};
 use crate::utils::translate_number_type;
 use crate::TransactionFor;
 use domain_block_preprocessor::preprocessor::SystemDomainBlockPreprocessor;
-use domain_block_preprocessor::state_root_extractor::StateRootExtractorWithSystemDomainClient;
+use domain_block_preprocessor::runtime_api_full::RuntimeApiFull;
 use domain_runtime_primitives::{AccountId, DomainCoreApi};
 use sc_client_api::{AuxStore, BlockBackend, StateBackendFor};
 use sc_consensus::BlockImport;
@@ -28,7 +28,7 @@ where
     keystore: SyncCryptoStorePtr,
     system_domain_block_preprocessor: SystemDomainBlockPreprocessor<Block, PBlock, Client, PClient>,
     domain_block_processor: DomainBlockProcessor<Block, PBlock, Client, PClient, Backend, E>,
-    state_root_extractor: StateRootExtractorWithSystemDomainClient<Client>,
+    state_root_extractor: RuntimeApiFull<Client>,
 }
 
 impl<Block, PBlock, Client, PClient, Backend, E> Clone
@@ -94,7 +94,7 @@ where
             keystore,
             system_domain_block_preprocessor,
             domain_block_processor,
-            state_root_extractor: StateRootExtractorWithSystemDomainClient::new(client),
+            state_root_extractor: RuntimeApiFull::new(client),
         }
     }
